Allow Watching on "/" when using chroot
The Watchers are registered in the `watches` HashMap with key being the
path of interest. Once a watch has been registered for the root "/" and
an event has been triggered for it (i.e. `NodeChildrenChanged`), the call
to `cut_chroot` would be issued to identify the watches. However, the
current `cut_chroot` would return "", instead of the expected  key "/" that
has been already registered in the `watches` HashMap

Signed-off-by: Seb Ospina <kraige@gmail.com>